mcp/requirements.txt
python-dotenv
langchain_community
langchain_openai
langchain-core
langchain-chroma
langgraph
unstructured
langchain-text-splitters
uvicorn
fastapi
itsdangerous
pypdf
blockfrost-python
google-cloud-translate==2.0.1
cbor2